requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
compositeai==0.1.8
cryptography==43.0.1
distro==1.9.0
firebase-admin==6.5.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Levenshtein==0.26.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
<<<<<<< HEAD
=======
msgpack==1.1.0
>>>>>>> 4528470 (auth and db connection)
narwhals==1.9.0
numpy==2.1.1
openai==1.51.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
<<<<<<< HEAD
protobuf==5.28.2
pyarrow==17.0.0
=======
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
>>>>>>> 4528470 (auth and db connection)
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
<<<<<<< HEAD
=======
PyJWT==2.9.0
pyparsing==3.1.4
>>>>>>> 4528470 (auth and db connection)
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
pytz==2024.2
RapidFuzz==3.10.0
referencing==0.35.1
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
<<<<<<< HEAD
=======
rsa==4.9
>>>>>>> 4528470 (auth and db connection)
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
<<<<<<< HEAD
=======
uritemplate==4.1.1
>>>>>>> 4528470 (auth and db connection)
urllib3==2.2.3
